Safari-609.1.11.1/LayoutTests/pointerevents/pointer-id-crash.html

<!DOCTYPE html>
<html>
<head>
<meta charset=utf-8>
</head>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

'use strict';

test(() => {
    const element = document.body.appendChild(document.createElement("div"));
    for (let value of [0, Number.MAX_VALUE, Number.MIN_VALUE, Number.MAX_VALUE + 1, Number.MIN_VALUE - 1]) {
        element.hasPointerCapture(value);
        assert_throws("NotFoundError", () => element.setPointerCapture(value))
        assert_throws("NotFoundError", () => element.releasePointerCapture(value))
    }
    element.remove();
}, "Checking 0 can be used as a Pointer ID.");

</script>
</body>
</html>
